Phase 2 prep: uint8 obs buffer + re-enable persistence + 13x13 temp schedule
Land items 1-3 from PHASE_TWO_TODO.md so the 13x13 Phase 2 run fits
inside the 42.8 GB cgroup host:

1. Replay buffer stores obs as uint8 instead of float32 (lossless — every
   cell is 0/1). Cuts the 13x13 1M-sample buffer from 12.2 GB to 3.6 GB
   and drops PCIe H2D bandwidth for the obs portion 4x as a side effect.
   push/_store cast explicitly; load_from accepts legacy float32 saves.
   trainer.train_step casts to float AFTER the H2D copy so CUDA fuses
   the dtype promotion into the load.

2. Re-enable buffer.save_to() in train.py. The 4x smaller obs buffer
   means the np.savez transient also shrinks 4x and comfortably fits
   under the cgroup cap that forced us to disable persistence in Phase 1.

3. 13x13 preset: temperature_moves=40 (~1/3 of the ~120-move avg 13x13
   game, matching the tuned 9x9 run 2 ratio) and temperature_low=0.25
   to preserve runner-up move signal in training targets.

_test_correctness.py adds Stage 1c: float32->uint8->float32 round-trip
identity + save_to/load_from exact round-trip including dtype check,
per the "Verification before merging" section of PHASE_TWO_TODO.md.

+++ Season-5/AlphaZero/training/trainer.py
@@ -71,7 +71,10 @@ class Trainer:
             value_np = np.concatenate([val_m, val_a], axis=0)
         else:
             obs_np, policy_np, value_np = buffer.sample(self.cfg.batch_size)
-        obs = torch.from_numpy(obs_np).to(self.device)
+        # Buffer stores obs as uint8 (Phase 2 memory optimization).
+        # Cast to float AFTER the H2D copy so PCIe sees 1 byte/cell
+        # instead of 4 — CUDA fuses the dtype promotion into the load.
+        obs = torch.from_numpy(obs_np).to(self.device).float()
         target_policy = torch.from_numpy(policy_np).to(self.device)
         target_value = torch.from_numpy(value_np).to(self.device)
